moved the blob container functions to helpers and fixed a few issues introduced by the async/await inclusion of vocabulary
