+++ public/js/contribute/pad/render.js
@@ -958,8 +958,8 @@ async function addImgs(kwargs) {
   const fls = data.filter((d) => d.status === 200);
   // THE CONFIG WITH DATA HERE IS A BIT ANNOYING, BUT IT IS FOR CASES WITH A TEMPLATE, TO MAKE SURE THE VARS SET (e.g. THE INSTRUCTION) ARE MAINTAINED
   if (fls.length === 1) {
-    const afls = fls.map(() => {
-      return async (f) => {
+    const afls = fls.map((f) => {
+      return async () => {
         let datum = {};
         if (container) datum = container.datum();
         if (datum.type !== 'img') datum = { instruction: datum.instruction };
