+++ public/js/contribute/xlsx/load.js
@@ -37,8 +37,8 @@ function doLoad() {
 
   const searchForm = d3.select('body form#contribute');
   if (searchForm.node().attachEvent)
-    searchForm.node().attachEvent('submit', catchSubmit);
-  else searchForm.node().addEventListener('submit', catchSubmit);
+    searchForm.node().attachEvent('submit', async evt => await catchSubmit(evt));
+  else searchForm.node().addEventListener('submit', async evt => await catchSubmit(evt));
 }
 
 window.addEventListener('load', doLoad);
