+++ routes/helpers/blobcontainer/index.js
