+++ routes/request/img/buffer.js
@@ -2,7 +2,7 @@ const { app_storage } = include('config/')
 const fs = require('fs')
 const path = require('path')
 
-const { createContainer } = include('routes/upload/container_functions.js');
+const { blobContainer } = include('routes/helpers/')
 const { BlobServiceClient } = require('@azure/storage-blob');
 
 module.exports = async (req, res) => {
@@ -19,7 +19,7 @@ module.exports = async (req, res) => {
 		// ESTABLISH THE CONNECTION TO AZURE
 		const blobServiceClient = BlobServiceClient.fromConnectionString(process.env.AZURE_STORAGE_CONNECTION_STRING)
 		// FIND OR CREATE THE CONTAINER
-		containerClient = await createContainer(blobServiceClient)
+		containerClient = await blobContainer.createContainer(blobServiceClient)
 
 		const blobClient = containerClient.getBlockBlobClient(path.join(targetdir, name))
 		await blobClient.uploadData(buffer)
